+++ src/commands/cK02kringSetName.h
@@ -1,6 +1,6 @@
 void cK02kringSetName(){
     // Set DESC for Kring
-    int kringnr = commandParts.arg1.toInt();
+    size_t kringnr = commandParts.arg1.toInt();
     String name = commandParts.arg2;
     if (kringnr >= 0 && kringnr < RGB_NR_ITEMS) {
         kring[kringnr].desc = name;
