+++ src/commands/cK03kringStartStop.h
@@ -1,8 +1,8 @@
 void cK03kringStartStop(){
     // Set Start Stop led for Kring
-    int kringnr = commandParts.arg1.toInt();
-    int start = commandParts.arg2.toInt();
-    int stop = commandParts.arg3.toInt();
+    size_t kringnr = commandParts.arg1.toInt();
+    size_t start = commandParts.arg2.toInt();
+    size_t stop = commandParts.arg3.toInt();
     if (kringnr >= 0 && kringnr < RGB_NR_ITEMS) {
         kring[kringnr].startrgb = start;
         kring[kringnr].stoprgb = stop;
