+++ src/commands/cK08kringTonToff.h
@@ -1,7 +1,7 @@
 void cK08kringTonToff(){
     // Set effect and time effect for Kring
-    int kringnr = commandParts.arg1.toInt();
-    int effect = commandParts.arg2.toInt();
+    size_t kringnr = commandParts.arg1.toInt();
+    size_t effect = commandParts.arg2.toInt();
     unsigned long timeeffect = commandParts.arg3.toDouble();
     if (kringnr >= 0 && kringnr < RGB_NR_ITEMS) {
         kring[kringnr].effect = effect;
