+++ src/commands/cK10kringSeed.h
@@ -1,7 +1,7 @@
 void cK10kringSeed(){
     // Set JPEG for Kring
-    int kringnr = commandParts.arg1.toInt();
-    int seed = commandParts.arg2.toInt();
+    size_t kringnr = commandParts.arg1.toInt();
+    size_t seed = commandParts.arg2.toInt();
     if (kringnr >= 0 && kringnr < RGB_NR_ITEMS) {
         kring[kringnr].seed = seed;
     } else {
