+++ src/commands/cL21ledLoadSetting.h
@@ -1,6 +1,6 @@
 void cL21ledLoadSetting(){
     // Load settings for a specific led
-    int lednr = commandParts.arg1.toInt();
+    size_t lednr = commandParts.arg1.toInt();
     String animatie = commandParts.arg2;
 
     if (lednr >= 0 && lednr < LED_NR_ITEMS) {
