+++ src/spiffs/fileLedLoadSettings.h
@@ -9,7 +9,7 @@ String  fileLedLoadSettings() {
 
   if (!SPIFFS.exists(filename)) {
     msgAnswer = "Led config file does not exist, using defaults";
-    for(int lednr=0; lednr<LED_NR_ITEMS; lednr++) {
+    for(size_t lednr=0; lednr<LED_NR_ITEMS; lednr++) {
       ledsettings[lednr].leddesc          = LEDACTIVE_DEFAULT.desc + String(lednr);
       ledsettings[lednr].ledimage         = LEDACTIVE_DEFAULT.image;
       ledsettings[lednr].anim[0].animdesc = LEDACTIVE_DEFAULT.animatie;
@@ -23,7 +23,7 @@ String  fileLedLoadSettings() {
   }
   else {
     String file_content = readFile(SPIFFS, filename);
-    int config_file_size = file_content.length();
+    size_t config_file_size = file_content.length();
     Println(" size: " + String(config_file_size));
 
     JsonDocument doc;
@@ -33,23 +33,23 @@ String  fileLedLoadSettings() {
       msgAnswer ="Error interpreting config file Led";
     }
 
-    int regelcounter = 0;
+    size_t regelcounter = 0;
     JsonArray jledarr = doc.as<JsonArray>();
     Println("aantal leds: " + String(jledarr.size()));
-    for (int lednr=0; lednr<jledarr.size(); lednr++) {
+    for (size_t lednr=0; lednr<jledarr.size(); lednr++) {
       ledsettings[lednr].leddesc          = jledarr[lednr]["leddesc"].as<String>();
       ledsettings[lednr].ledimage         = jledarr[lednr]["ledimage"].as<String>();
       Print(String(lednr)+" desc: " + jledarr[lednr]["leddesc"].as<String>());
 
       JsonArray janimarr = jledarr[lednr]["anim"].as<JsonArray>();
       Println(" aantal anims: " + String(janimarr.size()));
-      for (int animnr=0; animnr<janimarr.size(); animnr++) {
+      for (size_t animnr=0; animnr<janimarr.size(); animnr++) {
         ledsettings[lednr].anim[animnr].animdesc    = janimarr[animnr]["animdesc"].as<String>();
-        ledsettings[lednr].anim[animnr].bri         = janimarr[animnr]["bri"].as<int>();
-        ledsettings[lednr].anim[animnr].timeon      = janimarr[animnr]["timeon"].as<int>();
-        ledsettings[lednr].anim[animnr].timeoff     = janimarr[animnr]["timeoff"].as<int>();
-        ledsettings[lednr].anim[animnr].timeeffect  = janimarr[animnr]["timeeffect"].as<int>();
-        ledsettings[lednr].anim[animnr].effect      = janimarr[animnr]["effect"].as<int>();
+        ledsettings[lednr].anim[animnr].bri         = janimarr[animnr]["bri"].as<size_t>();
+        ledsettings[lednr].anim[animnr].timeon      = janimarr[animnr]["timeon"].as<size_t>();
+        ledsettings[lednr].anim[animnr].timeoff     = janimarr[animnr]["timeoff"].as<size_t>();
+        ledsettings[lednr].anim[animnr].timeeffect  = janimarr[animnr]["timeeffect"].as<size_t>();
+        ledsettings[lednr].anim[animnr].effect      = janimarr[animnr]["effect"].as<size_t>();
         regelcounter++; 
       }
