+++ src/spiffs/fileRgbLoadSettings.h
@@ -9,7 +9,7 @@ String  fileRgbLoadSettings() {
 
   if (!SPIFFS.exists(filename)) {
     msgAnswer = "Rgb config file does not exist, using defaults";
-    for(int rgbnr=0; rgbnr<RGB_NR_ITEMS; rgbnr++) {
+    for(size_t rgbnr=0; rgbnr<RGB_NR_ITEMS; rgbnr++) {
       rgbsettings[rgbnr].rgbdesc          = RGBACTIVE_DEFAULT.desc + String(rgbnr);
       rgbsettings[rgbnr].rgbimage         = RGBACTIVE_DEFAULT.image;
       rgbsettings[rgbnr].startrgb         = RGBACTIVE_DEFAULT.startrgb;
@@ -37,7 +37,7 @@ String  fileRgbLoadSettings() {
   }
   else {
     String file_content = readFile(SPIFFS, filename);
-    int config_file_size = file_content.length();
+    size_t config_file_size = file_content.length();
     Println(" size: " + String(config_file_size));
 
     JsonDocument doc;
@@ -47,38 +47,38 @@ String  fileRgbLoadSettings() {
       msgAnswer ="Error interpreting config file Rgb";
     }
 
-    int regelcounter = 0;
+    size_t regelcounter = 0;
     JsonArray jrgbarr = doc.as<JsonArray>();
     Println("aantal rgbs: " + String(jrgbarr.size()));
-    for (int rgbnr=0; rgbnr<jrgbarr.size(); rgbnr++) {
+    for (size_t rgbnr=0; rgbnr<jrgbarr.size(); rgbnr++) {
       rgbsettings[rgbnr].rgbdesc          = jrgbarr[rgbnr]["rgbdesc"].as<String>();
       rgbsettings[rgbnr].rgbimage         = jrgbarr[rgbnr]["rgbimage"].as<String>();
-      rgbsettings[rgbnr].startrgb         = jrgbarr[rgbnr]["startrgb"].as<int>();
-      rgbsettings[rgbnr].stoprgb          = jrgbarr[rgbnr]["stoprgb"].as<int>();
+      rgbsettings[rgbnr].startrgb         = jrgbarr[rgbnr]["startrgb"].as<size_t>();
+      rgbsettings[rgbnr].stoprgb          = jrgbarr[rgbnr]["stoprgb"].as<size_t>();
       Print(String(rgbnr)+" desc: " + jrgbarr[rgbnr]["rgbdesc"].as<String>());
 
       JsonArray janimarr = jrgbarr[rgbnr]["anim"].as<JsonArray>();
       Println(" aantal anims: " + String(janimarr.size()));
-      for (int animnr=0; animnr<janimarr.size(); animnr++) {
+      for (size_t animnr=0; animnr<janimarr.size(); animnr++) {
         rgbsettings[rgbnr].anim[animnr].animdesc    = janimarr[animnr]["animdesc"].as<String>();
-        rgbsettings[rgbnr].anim[animnr].hue1        = janimarr[animnr]["hue1"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].sat1        = janimarr[animnr]["sat1"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].bri1        = janimarr[animnr]["bri1"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].hue2        = janimarr[animnr]["hue2"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].sat2        = janimarr[animnr]["sat2"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].bri2        = janimarr[animnr]["bri2"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].hue3        = janimarr[animnr]["hue3"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].sat3        = janimarr[animnr]["sat3"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].bri3        = janimarr[animnr]["bri3"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].hue4        = janimarr[animnr]["hue4"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].sat4        = janimarr[animnr]["sat4"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].bri4        = janimarr[animnr]["bri4"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].every       = janimarr[animnr]["every"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].timeon      = janimarr[animnr]["timeon"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].timeoff     = janimarr[animnr]["timeoff"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].timeeffect  = janimarr[animnr]["timeeffect"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].seed        = janimarr[animnr]["seed"].as<int>();
-        rgbsettings[rgbnr].anim[animnr].effect      = janimarr[animnr]["effect"].as<int>();
+        rgbsettings[rgbnr].anim[animnr].hue1        = janimarr[animnr]["hue1"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].sat1        = janimarr[animnr]["sat1"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].bri1        = janimarr[animnr]["bri1"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].hue2        = janimarr[animnr]["hue2"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].sat2        = janimarr[animnr]["sat2"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].bri2        = janimarr[animnr]["bri2"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].hue3        = janimarr[animnr]["hue3"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].sat3        = janimarr[animnr]["sat3"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].bri3        = janimarr[animnr]["bri3"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].hue4        = janimarr[animnr]["hue4"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].sat4        = janimarr[animnr]["sat4"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].bri4        = janimarr[animnr]["bri4"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].every       = janimarr[animnr]["every"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].timeon      = janimarr[animnr]["timeon"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].timeoff     = janimarr[animnr]["timeoff"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].timeeffect  = janimarr[animnr]["timeeffect"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].seed        = janimarr[animnr]["seed"].as<size_t>();
+        rgbsettings[rgbnr].anim[animnr].effect      = janimarr[animnr]["effect"].as<size_t>();
         regelcounter++; 
       }
